refactor: fold ensure-propose-result into propose-grammar-update --ensure-result
Drops the standalone script/test in favor of a flag on the main tool, reusing
load_result_row's classification instead of a second hand-rolled JSON check.


# Conflicts:
#	Makefile
#	scripts/test_propose_grammar_update.py
