+++ .github/workflows/propose-grammar-updates.yml
@@ -228,7 +228,7 @@ jobs:
       # Fallback if tee left an empty/invalid artifact.
       - name: Ensure result JSON
         if: always()
-        run: python scripts/ensure-propose-result "result-${{ matrix.language }}.json"
+        run: python scripts/propose-grammar-update --ensure-result "result-${{ matrix.language }}.json"
       - uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
         if: always()
         with:
