+++ Makefile
@@ -52,7 +52,7 @@ test: build
 # Uses the 'pytest' on PATH if present, otherwise falls back to 'python3 -m pytest'.
 # scripts/test_propose_grammar_update.py MUST run as a separate invocation:
 # it's a uv script with its own pinned deps.
-PYTHON_TESTS = lang/test_grammar_ts_version.py lang/test_ts_versions.py lang/test_abi15_gating.py scripts/test_update_grammar.py scripts/test_cursor_agent_runner.py scripts/test_ensure_propose_result.py
+PYTHON_TESTS = lang/test_grammar_ts_version.py lang/test_ts_versions.py lang/test_abi15_gating.py scripts/test_update_grammar.py scripts/test_cursor_agent_runner.py
 .PHONY: test-python
 test-python:
 	@if command -v pytest >/dev/null 2>&1; then \
